+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Compiler errors when safe mode (`GLTFAST_SAFE` scripting define) is enabled.
 - Compiler error with High Definition Render Pipeline version 17 (2023.3)
 - Removed usage of obsolete APIs in High Definition Render Pipeline version 17 (2023.3)
+- (Export) Area light's range value is exported accurately (as shown in the inspector).
 
 ## [6.2.0] - 2024-01-29